[Test]Update TDD test to use BigDecimal
After converting floats to BigDecimal, i also have to update the TDDD
tests to be compataibale with BigDecimal to.

+++ calculator/src/test/java/calculator/TestCalcModel.java
@@ -3,6 +3,7 @@ package calculator;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import enums.OpType;
+import java.math.BigDecimal;
 import model.CalcModel;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,14 +20,14 @@ class TestCalcModel {
   @Test
   void testStandardEvaluate() {
     calc.setState(OpType.INFIX);
-    assertEquals(6.0f, calc.evaluate("3 + 3"));
+    assertEquals(BigDecimal.valueOf(6), calc.evaluate("3 + 3"));
   }
   
   //Change state to postfix then perform evaluation
   @Test
   void testPostfixEvaluate() {
     calc.setState(OpType.POSTFIX);
-    assertEquals(6.0f, calc.evaluate("3 3 +"));
+    assertEquals(BigDecimal.valueOf(6), calc.evaluate("3 3 +"));
   }
 
 }
